修改更新

+++ src/lib/drivers/device/integrator.h
@@ -67,7 +67,7 @@ public:
 	 * @return		true if putting the item triggered an integral reset and the integral should be
 	 *			published.
 	 */
-	bool put(uint64_t timestamp, matrix::Vector3f &val, matrix::Vector3f &integral, uint64_t &integral_dt);
+	bool put(uint64_t timestamp, matrix::Vector3f &val, matrix::Vector3f &integral, uint32_t &integral_dt);
 
 	/**
 	 * Put an item into the integral but provide an interval instead of a timestamp.
@@ -81,8 +81,7 @@ public:
 	 * @return		true if putting the item triggered an integral reset and the integral should be
 	 *			published.
 	 */
-	bool put_with_interval(unsigned interval_us, matrix::Vector3f &val, matrix::Vector3f &integral,
-			       uint64_t &integral_dt);
+	bool put_with_interval(unsigned interval_us, matrix::Vector3f &val, matrix::Vector3f &integral, uint32_t &integral_dt);
 
 	/**
 	 * Get the current integral and reset the integrator if needed.
@@ -91,7 +90,7 @@ public:
 	 * @param integral_dt	Get the dt in us of the current integration (only if reset).
 	 * @return		the integral since the last read-reset
 	 */
-	matrix::Vector3f	get(bool reset, uint64_t &integral_dt);
+	matrix::Vector3f	get(bool reset, uint32_t &integral_dt);
 
 
 	/**
@@ -103,7 +102,7 @@ public:
 	 * @param filtered_val	The integral differentiated by the integration time.
 	 * @return		the integral since the last read-reset
 	 */
-	matrix::Vector3f	get_and_filtered(bool reset, uint64_t &integral_dt, matrix::Vector3f &filtered_val);
+	matrix::Vector3f	get_and_filtered(bool reset, uint32_t &integral_dt, matrix::Vector3f &filtered_val);
 
 
 	/**
@@ -132,5 +131,5 @@ private:
 	 *
 	 * @param integral_dt	Get the dt in us of the current integration.
 	 */
-	void _reset(uint64_t &integral_dt);
+	void _reset(uint32_t &integral_dt);
 };
